Split container methods
