+++ src/Codeception/Module/Laravel.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Codeception\Module;
 
-use Closure;
 use Codeception\Configuration;
 use Codeception\Exception\ModuleConfigException;
 use Codeception\Lib\Connector\Laravel as LaravelConnector;
@@ -14,6 +13,7 @@ use Codeception\Lib\Interfaces\PartedModule;
 use Codeception\Lib\ModuleContainer;
 use Codeception\Module\Laravel\InteractsWithAuthentication;
 use Codeception\Module\Laravel\InteractsWithConsole;
+use Codeception\Module\Laravel\InteractsWithContainer;
 use Codeception\Subscriber\ErrorHandler;
 use Codeception\TestInterface;
 use Codeception\Util\ReflectionHelper;
@@ -131,6 +131,7 @@ class Laravel extends Framework implements ActiveRecord, PartedModule
 {
     use InteractsWithAuthentication;
     use InteractsWithConsole;
+    use InteractsWithContainer;
 
     /**
      * @var Application
@@ -290,24 +291,6 @@ class Laravel extends Framework implements ActiveRecord, PartedModule
     {
         $handler = new ErrorHandler();
         set_error_handler([$handler, 'errorHandler']);
-    }
-
-    /**
-     * Provides access the Laravel application object.
-     *
-     * @return \Illuminate\Contracts\Foundation\Application
-     */
-    public function getApplication()
-    {
-        return $this->app;
-    }
-
-    /**
-     * @param \Illuminate\Contracts\Foundation\Application $app
-     */
-    public function setApplication($app): void
-    {
-        $this->app = $app;
     }
 
     /**
@@ -764,32 +747,6 @@ class Laravel extends Framework implements ActiveRecord, PartedModule
     }
 
     /**
-     * Return an instance of a class from the Laravel service container.
-     * (https://laravel.com/docs/master/container)
-     *
-     * ``` php
-     * <?php
-     * // In Laravel
-     * App::bind('foo', function($app) {
-     *     return new FooBar;
-     * });
-     *
-     * // Then in test
-     * $service = $I->grabService('foo');
-     *
-     * // Will return an instance of FooBar, also works for singletons.
-     * ```
-     *
-     * @param string $class
-     * @return mixed
-     */
-    public function grabService(string $class)
-    {
-        return $this->app[$class];
-    }
-
-
-    /**
      * Inserts record into the database.
      * If you pass the name of a database table as the first argument, this method returns an integer ID.
      * You can also pass the class name of an Eloquent model, in that case this method returns an Eloquent model.
@@ -1231,109 +1188,4 @@ class Laravel extends Framework implements ActiveRecord, PartedModule
         return $query;
     }
 
-    /**
-     * Add a binding to the Laravel service container.
-     * (https://laravel.com/docs/master/container)
-     *
-     * ``` php
-     * <?php
-     * $I->haveBinding('My\Interface', 'My\Implementation');
-     * ```
-     *
-     * @param string $abstract
-     * @param Closure|string|null $concrete
-     * @param bool $shared
-     */
-    public function haveBinding(string $abstract, $concrete = null, bool $shared = false): void
-    {
-        $this->client->haveBinding($abstract, $concrete, $shared);
-    }
-
-    /**
-     * Add a singleton binding to the Laravel service container.
-     * (https://laravel.com/docs/master/container)
-     *
-     * ``` php
-     * <?php
-     * $I->haveSingleton('App\MyInterface', 'App\MySingleton');
-     * ```
-     *
-     * @param string $abstract
-     * @param Closure|string|null $concrete
-     */
-    public function haveSingleton(string $abstract, $concrete): void
-    {
-        $this->client->haveBinding($abstract, $concrete, true);
-    }
-
-    /**
-     * Add a contextual binding to the Laravel service container.
-     * (https://laravel.com/docs/master/container)
-     *
-     * ``` php
-     * <?php
-     * $I->haveContextualBinding('My\Class', '$variable', 'value');
-     *
-     * // This is similar to the following in your Laravel application
-     * $app->when('My\Class')
-     *     ->needs('$variable')
-     *     ->give('value');
-     * ```
-     *
-     * @param string $concrete
-     * @param string $abstract
-     * @param Closure|string $implementation
-     */
-    public function haveContextualBinding(string $concrete, string $abstract, $implementation): void
-    {
-        $this->client->haveContextualBinding($concrete, $abstract, $implementation);
-    }
-
-    /**
-     * Add an instance binding to the Laravel service container.
-     * (https://laravel.com/docs/master/container)
-     *
-     * ``` php
-     * <?php
-     * $I->haveInstance('App\MyClass', new App\MyClass());
-     * ```
-     *
-     * @param string $abstract
-     * @param mixed $instance
-     */
-    public function haveInstance(string $abstract, $instance): void
-    {
-        $this->client->haveInstance($abstract, $instance);
-    }
-
-    /**
-     * Register a handler than can be used to modify the Laravel application object after it is initialized.
-     * The Laravel application object will be passed as an argument to the handler.
-     *
-     * ``` php
-     * <?php
-     * $I->haveApplicationHandler(function($app) {
-     *     $app->make('config')->set(['test_value' => '10']);
-     * });
-     * ```
-     *
-     * @param callable $handler
-     */
-    public function haveApplicationHandler(callable $handler): void
-    {
-        $this->client->haveApplicationHandler($handler);
-    }
-
-    /**
-     * Clear the registered application handlers.
-     *
-     * ``` php
-     * <?php
-     * $I->clearApplicationHandlers();
-     * ```
-     */
-    public function clearApplicationHandlers(): void
-    {
-        $this->client->clearApplicationHandlers();
-    }
 }
